Refined the GUI and figured out how to do raycast, but it's not used by an actual AI thinker yet.

+++ Assets/Resources/Scripts/UI/VitalsUI.cs
@@ -36,7 +36,7 @@ public class VitalsUI : MonoBehaviour {
 		
 	}
 
-    public void InitializeVitals(int maxHP, int maxMP){
+    public void InitializeVitals(float maxHP, float maxMP){
         //HPOuterBarTransform.rect.Set();
         HPOuterBarTransform.sizeDelta = new Vector2(maxHP * scale+outlineOffset, HPOuterBarTransform.sizeDelta.y);
         HPBackgroundBarTransform.sizeDelta = new Vector2(maxHP * scale, HPBackgroundBarTransform.sizeDelta.y);
